replaced double values to decimal

+++ SpecFlowProject/Client/Constants/ConstantsFoodCosts.cs
@@ -2,7 +2,7 @@ namespace SpecFlowProject.Client.Constants;
 
 public class ConstantsFoodCosts
 {
-    public const double Starters = 4.00;
-    public const double Mains = 7.00;
-    public const double Drinks = 2.50;
+    public const decimal Starters = 4;
+    public const decimal Mains = 7;
+    public const decimal Drinks = (decimal)2.5;
 }
